+++ packages/orq-rc/src/models/components/conversationresponse.ts
@@ -128,7 +128,7 @@ export const ConversationResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kdyzdpq9j2skkx1tnv0k65s3"),
+  _id: z.string().default("conv_01ke6fm0yg2dm26zz8a7nr60jp"),
   entityId: z.string(),
   kind: ConversationResponseKind$inboundSchema,
   displayName: z.string(),
@@ -161,7 +161,7 @@ export const ConversationResponse$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ConversationResponse
 > = z.object({
-  id: z.string().default("conv_01kdyzdpq9j2skkx1tnv0k65s3"),
+  id: z.string().default("conv_01ke6fm0yg2dm26zz8a7nr60jp"),
   entityId: z.string(),
   kind: ConversationResponseKind$outboundSchema,
   displayName: z.string(),