+++ packages/orq-rc/RELEASES.md
@@ -5689,3 +5689,13 @@ Based on:
 - [typescript v4.2.0-rc.8] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.8 - packages/orq-rc
+
+## 2026-01-05 07:08:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.680.12 (2.788.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.9] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.9 - packages/orq-rc